+++ src/main/kotlin/org/axonframework/intellij/ide/plugin/inspections/aggregate/JavaAggregateMemberRoutingKeyInspection.kt
@@ -14,22 +14,6 @@
  *  limitations under the License.
  */
 
-/*
- *  Copyright (c) 2022. Axon Framework
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
 package org.axonframework.intellij.ide.plugin.inspections.aggregate
 
 import com.intellij.codeInspection.AbstractBaseJavaLocalInspectionTool
@@ -51,8 +35,8 @@ class JavaAggregateMemberRoutingKeyInspection : AbstractBaseJavaLocalInspectionT
         isOnTheFly: Boolean
     ): Array<ProblemDescriptor>? {
         val name = aClass.qualifiedName ?: return null
-        val member = aClass.aggregateResolver().getMemberForName(name) ?: return null
-        return member.children.filter { it.isCollection && !it.member.entityIdPresent }.map { child ->
+        val entity = aClass.aggregateResolver().getEntityByName(name) ?: return null
+        return entity.members.filter { it.isCollection && it.member.routingKey == null }.map { child ->
             val classField = aClass.fields.first { it.name == child.fieldName }
             manager.createProblemDescriptor(
                 classField,